libnvme/mi: widen MPRT before multiplying by 100
le16_to_cpu(msg->mprt) is at most 16 bits wide; multiplying by 100
can reach 6,553,500, relying on implicit promotion to a wide enough
int for the result to fit. Cast to unsigned int explicitly so the
multiplication's width doesn't depend on int's size.

+++ libnvme/src/nvme/mi-mctp.c
@@ -239,7 +239,7 @@ static bool libnvme_mi_mctp_resp_is_mpr(void *buf, size_t len,
 		return false;
 
 	if (mpr_time)
-		*mpr_time = le16_to_cpu(msg->mprt) * 100;
+		*mpr_time = (unsigned int)le16_to_cpu(msg->mprt) * 100;
 
 	return true;
 }
